fix: Fixing error return for removeDocumento

+++ src/reunioes/reunioes.service.ts
@@ -46,21 +46,30 @@ export class ReunioesService {
   }
 
   async removeDocumento(id: string) {
+    const documento = await this.prisma.documento.findUnique({
+      where: {
+        id_documento: id,
+      },
+    });
+  
+    if (!documento) {
+      throw new NotFoundException('Documento não encontrado.');
+    }
+  
     return await this.prisma.documento.delete({
       where: {
-        id_documento: id
-      }
+        id_documento: id,
+      },
     }).then((documento) => {
       return {
         message: "Documento deletado com sucesso",
-        documento: documento
-      }
-    }
-    ).catch((err) => {
+        documento: documento,
+      };
+    }).catch((err) => {
       console.log(err);
       throw new UnprocessableEntityException('Erro ao deletar documento.');
     });
-  }
+  }  
 
   async getReunioes(idusuario: string, nivel_acesso: NivelAcesso) {
     let idKey;
